% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{exampleData500}
\alias{exampleData500}
\title{A simulated toy dataset that includes 2 data matrices (from 2 related
tasks).}
\format{A list of 2 matrices of size [1:500, 1:500] and [1:500, 1:500]}
\usage{
data(exampleData500)
}
\description{
A simulated toy dataset that includes 2 data matrices (from 2 related
tasks). Each data matrix is about 500 features observed in 1000 samples. The
two data matrices are about exactly the same set of 500 features. This
multi-task dataset is generated from two related random graphs. Please run
demo(diffee) to learn the basic functions provided by this package.  For
further details, please read the original paper:
<http://link.springer.com/article/10.1007/s10994-017-5635-7>.
}
\keyword{datasets}
